% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.conindex}
\alias{coef.conindex}
\title{Extract the concentration index estimate}
\usage{
\method{coef}{conindex}(object, ...)
}
\arguments{
\item{object}{A \code{conindex} object.}

\item{...}{Ignored.}
}
\value{
The index value (numeric scalar).
}
\description{
Extract the concentration index estimate
}
