% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.conindex}
\alias{print.conindex}
\alias{print.conindex_compare}
\title{Print a concentration index}
\usage{
\method{print}{conindex}(x, digits = 7, ...)

\method{print}{conindex_compare}(x, digits = 7, ...)
}
\arguments{
\item{x}{A \code{conindex} object.}

\item{digits}{Number of significant digits.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a concentration index
}
