% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by.R
\name{print.conindex_by}
\alias{print.conindex_by}
\title{Print by-group concentration indices}
\usage{
\method{print}{conindex_by}(x, ...)
}
\arguments{
\item{x}{A \code{conindex_by} object.}

\item{...}{Passed to \code{\link[=print.conindex]{print.conindex()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Print by-group concentration indices
}
