% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.conindex}
\alias{summary.conindex}
\title{Summarise a concentration index}
\usage{
\method{summary}{conindex}(object, ...)
}
\arguments{
\item{object}{A \code{conindex} object.}

\item{...}{Ignored.}
}
\value{
\code{object}, invisibly (prints a summary).
}
\description{
Summarise a concentration index
}
